(#97) moving DTOs to the services
