+++ storage/src/main/java/org/zalando/apidiscovery/storage/api/resource/ApiResourceController.java
@@ -10,14 +10,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
-import org.zalando.apidiscovery.storage.api.resource.dto.ApiDefinitionDto;
-import org.zalando.apidiscovery.storage.api.resource.dto.ApiDto;
+import org.zalando.apidiscovery.storage.api.service.dto.ApiDefinitionDto;
+import org.zalando.apidiscovery.storage.api.service.dto.ApiDto;
 import org.zalando.apidiscovery.storage.api.domain.ApiLifecycleState;
-import org.zalando.apidiscovery.storage.api.resource.dto.ApiListDto;
-import org.zalando.apidiscovery.storage.api.resource.dto.DeploymentDto;
-import org.zalando.apidiscovery.storage.api.resource.dto.DeploymentsDto;
-import org.zalando.apidiscovery.storage.api.resource.dto.VersionListDto;
-import org.zalando.apidiscovery.storage.api.resource.dto.VersionsDto;
+import org.zalando.apidiscovery.storage.api.service.dto.ApiListDto;
+import org.zalando.apidiscovery.storage.api.service.dto.DeploymentDto;
+import org.zalando.apidiscovery.storage.api.service.dto.DeploymentsDto;
+import org.zalando.apidiscovery.storage.api.service.dto.VersionListDto;
+import org.zalando.apidiscovery.storage.api.service.dto.VersionsDto;
 import org.zalando.apidiscovery.storage.api.service.ApiService;
 
 import java.util.List;
