+++ storage/src/main/java/org/zalando/apidiscovery/storage/api/service/ApiService.java
@@ -3,11 +3,11 @@ package org.zalando.apidiscovery.storage.api.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.zalando.apidiscovery.storage.api.domain.ApiLifecycleState;
-import org.zalando.apidiscovery.storage.api.resource.dto.ApiDefinitionDto;
-import org.zalando.apidiscovery.storage.api.resource.dto.ApiDto;
-import org.zalando.apidiscovery.storage.api.resource.dto.ApplicationDto;
-import org.zalando.apidiscovery.storage.api.resource.dto.DeploymentDto;
-import org.zalando.apidiscovery.storage.api.resource.dto.VersionsDto;
+import org.zalando.apidiscovery.storage.api.service.dto.ApiDefinitionDto;
+import org.zalando.apidiscovery.storage.api.service.dto.ApiDto;
+import org.zalando.apidiscovery.storage.api.service.dto.ApplicationDto;
+import org.zalando.apidiscovery.storage.api.service.dto.DeploymentDto;
+import org.zalando.apidiscovery.storage.api.service.dto.VersionsDto;
 import org.zalando.apidiscovery.storage.api.domain.util.ApiEntityToApiDefinitionConverter;
 import org.zalando.apidiscovery.storage.api.repository.ApiDeploymentEntity;
 import org.zalando.apidiscovery.storage.api.repository.ApiEntity;
