+++ storage/src/main/java/org/zalando/apidiscovery/storage/api/service/dto/ApiDefinitionDto.java
@@ -1,4 +1,4 @@
-package org.zalando.apidiscovery.storage.api.resource.dto;
+package org.zalando.apidiscovery.storage.api.service.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
