+++ storage/src/main/java/org/zalando/apidiscovery/storage/api/service/dto/ApplicationsDto.java
@@ -1,4 +1,4 @@
-package org.zalando.apidiscovery.storage.api.resource.dto;
+package org.zalando.apidiscovery.storage.api.service.dto;
 
 import java.util.ArrayList;
 import java.util.List;
