+++ storage/src/test/java/org/zalando/apidiscovery/storage/api/domain/util/ApiEntityToApiDefinitionConverterTest.java
@@ -1,12 +1,12 @@
 package org.zalando.apidiscovery.storage.api.domain.util;
 
 import org.junit.Test;
-import org.zalando.apidiscovery.storage.api.resource.dto.ApiDefinitionDto;
+import org.zalando.apidiscovery.storage.api.service.dto.ApiDefinitionDto;
 import org.zalando.apidiscovery.storage.api.DomainObjectGen;
 import org.zalando.apidiscovery.storage.api.repository.ApiDeploymentEntity;
 import org.zalando.apidiscovery.storage.api.repository.ApiEntity;
 import org.zalando.apidiscovery.storage.api.repository.ApplicationEntity;
-import org.zalando.apidiscovery.storage.api.resource.dto.DeploymentLinkDto;
+import org.zalando.apidiscovery.storage.api.service.dto.DeploymentLinkDto;
 
 import static java.util.Arrays.asList;
 import static org.assertj.core.api.Assertions.assertThat;
