+++ storage/src/test/java/org/zalando/apidiscovery/storage/api/domain/util/ApiEntityToVersionConverterTest.java
@@ -2,7 +2,7 @@ package org.zalando.apidiscovery.storage.api.domain.util;
 
 import org.junit.Test;
 import org.zalando.apidiscovery.storage.api.repository.ApiEntity;
-import org.zalando.apidiscovery.storage.api.resource.dto.VersionsDto;
+import org.zalando.apidiscovery.storage.api.service.dto.VersionsDto;
 
 import java.util.List;
