+++ storage/src/test/java/org/zalando/apidiscovery/storage/api/service/dto/ApiDtoJsonTest.java
@@ -1,4 +1,4 @@
-package org.zalando.apidiscovery.storage.api.resource.dto;
+package org.zalando.apidiscovery.storage.api.service.dto;
 
 
 import org.junit.Test;
@@ -7,7 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.zalando.apidiscovery.storage.api.resource.dto.ApiDto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.zalando.apidiscovery.storage.api.domain.ApiLifecycleState.ACTIVE;
